Changing filed_licence to field_license
Module cant be used without aditional changes to fileds, because in default package_show api, as defined in dkan_nkod.features.inc, the value for license link is set to [node:field-resources:Nth:field-license], but field in content type resource, as defined here, in dkan_nkod.features.field_instance.inc and dkan_nkod.info, has machine name "field_licence". So there is need for "unification".
Decided to change it here (and the other two files) in order to be consitent with the english (us) naming of machine names of fields.